requirements.txt
Django==3.2.5
djangorestframework==3.12.4
pytest==7.1.2
pytest-django==4.5.2
factory_boy==3.2.1
black==22.8.0
djangorestframework-simplejwt==5.2.0